+++ docs/CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - Add ``gtf-to-bed`` process
 - Add filter by sample to variant objects
 - Add additional parameters for ``calculate-bigwig`` process
+- Add ``samtools-mpileup-single`` process
 
 Changed
 -------